% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addrock.R
\name{addrock}
\alias{addrock}
\title{Medium resolution vector polygons of Antarctic rock outcrop - VERSION 7.3}
\usage{
addrock()
}
\value{
character string, a GDAL data source name for the Antarctic rock outcrop polygons
}
\description{
https://data.bas.ac.uk/full-record.php?id=GB/NERC/BAS/PDC/01409
}
\examples{
addrock()
}
