% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_meta_features.R
\name{build_meta_features}
\alias{build_meta_features}
\title{Feature reduction}
\usage{
build_meta_features(attrib_training, attrib_test, corrT = 0.6, top,
  scores)
}
\arguments{
\item{attrib_training}{Matrix of atrribute values with attribute names as row names and training sample names as column names.}

\item{attrib_test}{Matrix of atrribute values with attribute names as row names and test sample names as column names.}

\item{corrT}{Threshold to decide whether two features are correlated. Defaults to 0.6.}

\item{top}{Vector of selected attribute names.}

\item{scores}{Either score1 or score_combo. Vector of survival distance scores or combo with attribute names as names.}
}
\value{
Matrix of meta feature values with sample names as row names.
}
\description{
This function computes meta features from the best selected features by building a weighted mean of a group of pairwise (spearman) correlated features.
}
\examples{
build_meta_feature()
}
\keyword{correlation}
\keyword{feature}
\keyword{reduction}
