% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgaz.R
\name{CGAZ}
\alias{CGAZ}
\alias{CGAZ_sql}
\title{the geoBoundaries countries}
\usage{
CGAZ(old = FALSE)

CGAZ_sql(codes)
}
\arguments{
\item{old}{logical, return the old slow zipped shapefile or the new Parquet copy}

\item{codes}{a list of iso3 country codes, or country names (this is a bit sketchy)}
}
\value{
character string; \code{CGAZ()} returns a GDAL data source name and
\code{CGAZ_sql()} returns an SQL query string
}
\description{
CGAZ() returns the DSN, a shapefile of geo boundaries, CGAZ_sql() returns SQL suitable for use
with GDAL, for the names or codes of countries.
}
\examples{
CGAZ_sql(c("Australia", "New Zealand"))
CGAZ_sql(c("AUS", "NZL"))
## do something like gdal_raster_data(gebco(), target_res = 1,
##                                        options = c("-crop_to_cutline",
##                                        "-cutline", CGAZ(),
##                                         "-csql", CGAZ_sql(c("Australia", "New Zealand")) ))
}
