% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_notes.R
\name{class_notes}
\alias{class_notes}
\title{Class notes format (PDF)}
\usage{
class_notes(fig_width = 10, fig_height = 2.5, fig_crop = TRUE,
  dev = "pdf", highlight = "default", keep_tex = FALSE, includes = NULL,
  md_extensions = NULL, pandoc_args = NULL)
}
\description{
Class notes format (PDF)
}

