% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_scores.R
\name{combine_scores}
\alias{combine_scores}
\title{Score combination}
\usage{
combine_scores(scoreS, scoreC, weight)
}
\arguments{
\item{scoreS}{Vector of survival distance scores.}

\item{scoreC}{Vector of clinical independence scores.}

\item{weight}{Weight to apply to score 2.}
}
\value{
Vector of scores with attribute names as names. Higher is better.
}
\description{
This function combines the survival distance score (scoreS) and the clinical independence score (scoreC) with weighted standardization.
}
\examples{
combine_scores()
}
\keyword{feature}
\keyword{selection}
