% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_search.R
\name{course_search}
\alias{course_search}
\title{Search the online Smith College Course Catalog}
\usage{
course_search(
  year,
  semester,
  dept = "",
  instructor = "",
  instr_method = "",
  credits = "",
  course_number = "",
  course_keywords = "",
  course_levels = "",
  include_details = TRUE
)
}
\arguments{
\item{year}{Scalar character vector identifying the academic year you would like to search for classes during. AY 2022-23 is identified as 2023, AY 2021-22 is identified as 2022, etc.}

\item{semester}{Scalar character vector identifying you would like to search for classes during. Fall is identified as 01, J-Term is identified as 02, Spring is identified as 03.}

\item{dept}{Scalar character vector identifying a department or program by it's three letter abbreviation (SDS, ENG, MTH, etc.). Note that searching for a specific department can return courses that are cross-listed in that department.}

\item{instructor}{Scalar character vector describing the name of the instructor whose courses you want to search for.}

\item{instr_method}{Scalar character vector, either "Online" or "In-Person".}

\item{credits}{Scalar vector describing the number of credits you want the classes in the results to have. Can be numeric, but is coerced to character.}

\item{course_number}{Scalar vector describing the course number you are searching for (e.g., 192, 220). Can be numeric, but is coerced to character.}

\item{course_keywords}{Character vector identifying keyword to look for in course names and descriptions.}

\item{course_levels}{A numeric vector with a most 4 elements. Elements should be restricted to integer values 1, 2, 3 or 4 (indicating a search for courses at the 100, 200, 300, or 400 level).}

\item{include_details}{Logical indicating whether to include the expanded course details (e.g., prerequisites, curriculum distribution, etc.). This can be relatively slow, so for large searches it may be desirable to suppress this portion of the results.}
}
\value{
A data frame with up 19 variables:
\itemize{
\item Dept
\item Course.Num
\item Course.Sec
\item Title
\item Instructor
\item Waitlist
\item Credits
\item Course.Type
\item Grade.Mode
\item Coreq
\item Curriculum.Distribution
\item Time.Location
\item Max.Enrollment
\item Section.Enrollment
\item Waitlist.Count
\item Reserved.Seats
\item Enforced.Requirements
\item Instructional.Method
\item Description
}
}
\description{
Search the online Smith College Course Catalog
}
\examples{
\dontrun{
course_search(year = 2024, semester = "03", dept="SDS", course_levels = c(1, 2))
}

}
