% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{cull_html}
\alias{cull_html}
\alias{anonymize}
\alias{html_remove_byline}
\title{Make a gallery of HTML docs downloaded from Moodle}
\usage{
cull_html(from, to, ...)

anonymize(dir)

html_remove_byline(path, verbose = FALSE, ...)
}
\arguments{
\item{from}{directory of Moodle assignment}

\item{to}{directory of new gallery}

\item{...}{currently ignored}

\item{dir}{gallery directory}

\item{path}{file to be anonymized}

\item{verbose}{tell you what happened?}
}
\description{
Make a gallery of HTML docs downloaded from Moodle
}
\examples{
\dontrun{
my_dir <- "~/Dropbox/git/sds192/student_info/SDS192- 01_201801-Mini-project #4-265952/"
new_dir <- "~/Dropbox/git/sds192/www/gallery"
if (require(dplyr)) {
  my_dir \%>\%
    cull_html(to = new_dir) \%>\%
    anonymize()
}
}

}
