% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{dsn-sources}
\alias{dsn-sources}
\alias{wms_arcgis_mapserver_ESRI.WorldImagery_tms}
\alias{wms_arcgis_mapserver_tms}
\alias{wms_googlehybrid_tms}
\alias{wms_googleterrainstreets_tms}
\alias{wms_virtualearth}
\alias{wms_virtualearth_street}
\alias{wms_openstreetmap_tms}
\alias{wms_amazon_elevation}
\alias{wms_ESA_worldcover_2020_tms}
\alias{wms_mapbox_satellite}
\alias{wms_mapbox_terrain}
\alias{nasadem}
\alias{cop90}
\alias{cop30}
\alias{srtm15}
\title{Imagery online sources}
\usage{
wms_arcgis_mapserver_ESRI.WorldImagery_tms()

wms_arcgis_mapserver_tms()

wms_googlehybrid_tms()

wms_googleterrainstreets_tms()

wms_virtualearth()

wms_virtualearth_street()

wms_openstreetmap_tms()

wms_amazon_elevation()

wms_ESA_worldcover_2020_tms()

wms_mapbox_satellite()

wms_mapbox_terrain()

nasadem()

cop90()

cop30()

srtm15()
}
\description{
Raster and imagery online. Each function returns a GDAL-ready data source
name; for the \verb{wms_*} family that is a 'GDAL_WMS' XML recipe, stored as a
file in the package (see \code{dsn_list(kind = "xml_file")}). The two mapbox
functions return a template with a \verb{\%s} slot for an access token and are
not registry rows.
}
