% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esacci.R
\name{esacci_chlor_a}
\alias{esacci_chlor_a}
\title{ESACCI chlorophyll-a in monthly}
\usage{
esacci_chlor_a(time.resolution = "monthly")
}
\arguments{
\item{time.resolution}{monthly only for now}
}
\value{
data frame with date,source
}
\description{
ESACCI chlorophyll-a in monthly
}
\examples{
dsn <- esacci_chlor_a()
}
