% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadm.R
\name{gadm}
\alias{gadm}
\title{GADM administrative boundaries}
\usage{
gadm(vsi = TRUE)
}
\arguments{
\item{vsi}{include the 'vsicurl' prefix (\code{TRUE} is default)}
}
\value{
data source name for GADM 4.1 GeoParquet
}
\description{
GADM 4.1 whole-planet administrative areas as GeoParquet, hosted as an
sds release asset. The original single GeoPackage from geodata.ucdavis.edu
is available as \code{dsn("gadm_gpkg")}. GADM data is free for academic and
non-commercial use only, see gadm.org/license.
}
\examples{
gadm()
}
