% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{gedtm30_model}
\alias{gedtm30_model}
\title{GEDTM30 global 1-arc-second (~30m) Digital Terrain Model (DTM)}
\usage{
gedtm30_model(varname = "legendtm_rf_30m", vsi = TRUE)
}
\arguments{
\item{varname}{defaults to 'legendtm_rf_30m' (the elevation)}

\item{vsi}{include /vsicurl prefix}
}
\value{
for 'gedtm30_sources' dataframe of details about the files, for 'gedtm30_model' one of the variables source urls
}
\description{
GEDTM30 global 1-arc-second (~30m) Digital Terrain Model (DTM)
}
\examples{
gedtm30_model("hillshade_edtm")
}
\seealso{
\code{gedtm30()} which returns a single cog url just for the DEM
}
