% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score_c.R
\name{get_score_c}
\alias{get_score_c}
\title{Score c}
\usage{
get_score_c(attrib, clin)
}
\arguments{
\item{attrib}{Matrix of atrribute values with attribute names as row names and sample names as column names.}

\item{clin}{data.frame of clinical data with dummy variables for categorical variables and with sample names as row names and variable names as column names.}
}
\value{
Vector of scores with attribute names as names. Higher is better.
}
\description{
This function calculates the clinical independence score based on attribute values and clinical data.
}
\examples{
get_score_c()
}
\keyword{feature}
\keyword{selection}
