% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score_s.R
\name{get_score_s}
\alias{get_score_s}
\title{Score s}
\usage{
get_score_s(attrib, surv)
}
\arguments{
\item{attrib}{Matrix of atrribute values with attribute names as row names and sample names as column names.}

\item{surv}{data.frame of survival invormation with sample names as row names and column names 'overall.survival.indicator' and 'OVERALL.SURVIVAL'.}
}
\value{
Vector of scores with attribute names as names. Higher is better.
}
\description{
This function calculates the survival distance score based on attribute values and survival data.
}
\examples{
get_score_s()
}
\keyword{feature}
\keyword{selection}
