% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{ibcso}
\alias{ibcso}
\title{IBCOS source dsn}
\usage{
ibcso(vsi = TRUE, chart = FALSE)
}
\arguments{
\item{vsi}{include the 'vsicurl' prefix (\code{TRUE} is default)}

\item{chart}{the image or the data? set to TRUE for image (it's a PDF)}
}
\value{
character string, URL to online raster
}
\description{
A data source name to the IBCSO  elevation 'COG' GeoTIFF.
}
\details{
Currently at v2.
}
\examples{
ibcso()

}
