% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/india_data.R
\name{india_data}
\alias{india_data}
\title{Read India Data Sets}
\usage{
india_data(date_initial, date_final)
}
\description{
Read India Data Sets
}
\examples{

suppressMessages({
  dat = india_data("2020-04-01", "2020-06-01")
  dat = india_data("2020-04-01", "2020-07-01")
})


}
