% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listmap.R
\name{list_address}
\alias{list_address}
\title{Tasmania theLIST address query}
\usage{
list_address(address)
}
\arguments{
\item{address}{character vector of length 3: street number, street name,
and locality}
}
\value{
string URI for List service
}
\description{
Build a query URI for the Tasmanian theLIST address service.
}
\examples{
list_address(c(2862, "LYELL", "HAYES"))
}
