% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listmap.R
\name{list_parcel_shp}
\alias{list_parcel_shp}
\title{Cadastral parcel source}
\usage{
list_parcel_shp()
}
\value{
string URI for List shapefile
}
\description{
Cadastral parcel source
}
\examples{
list_parcel_shp()  ## read with terra::vect( ) or new(gdalraster::GDALVector, )
}
