% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapterhorn_elevation.R
\name{mapterhorn_elevation}
\alias{mapterhorn_elevation}
\title{Mapterhorn PMTiles raster elevation}
\usage{
mapterhorn_elevation(...)
}
\arguments{
\item{...}{unused}
}
\value{
VRT text string, useable by GDAL >= 3.14 with muparser support
}
\description{
Rendered calc VRT that unpacks planet.pmtiles from Mapterhorn, a large collection of
regional elevation sources encoded in Terrarium Terrain RGB '(R*256 + G + B/256) - 32768'
}
\examples{
writeLines(mapterhorn_elevation())
}
