% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_score.R
\name{meanRiskScoreFunction}
\alias{meanRiskScoreFunction}
\title{Risk Score Function}
\usage{
meanRiskScoreFunction(features_training, features_test,
  survObject_training, alpha = 0)
}
\arguments{
\item{features_training}{Matrix of selected (meta-)feature values with training sample names as row names.}

\item{features_test}{Matrix of selected (meta-)feature values with test sample names as row names.}

\item{survObject_training}{The survival data of the training sample in form of a survival object}

\item{alpha}{Alpha value which declares ridge regression (alpha = 0). For lasso set alpha = 1. Defaults to 0.}
}
\value{
List of one risk score per sample for training samples and test samples
}
\description{
This function computes the risk score for a ridge regression model with the selected features.
}
\examples{
meanRiskScoreFunction()
}
\keyword{prediction}
\keyword{risk}
