% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{mpc}
\alias{mpc}
\title{Microsoft Planetary Computer}
\usage{
mpc(
  collection = "sentinel-2-l2a",
  bbox = "146.5,-43.2,147.5,-42.2",
  datetime = Sys.Date() + c(-5, 1),
  asset = "",
  stacit = TRUE
)
}
\arguments{
\item{collection}{name of the collection e.g. "sentinel-2-l2a", "naip", or ""landsat-c2-l2"}

\item{bbox}{a string in the form 'xmin,ymin,xmax,ymax' where x,y are longitude and latitude values (OR a numeric extent xmin,xmax,ymin,ymax)}

\item{datetime}{a datetime of 1 or 2 values to give a range in time, if only one is given it is treated as an open interval to the present}

\item{asset}{name of the asset of interest, e.g. for "sentinel-2-l2a" there is "visual", "AOT", "B04" etc.}

\item{stacit}{logical, treat this as a 'GDAL STACIT' source, or just a generic MPC query, 'TRUE' by default}
}
\value{
a string, a data source name for 'GDAL'
}
\description{
Function \code{mpc_stacit()} will return a full data source name for a Spatio-Temporal Asset Catalog ('STAC').
}
\details{
Each argument has a default, use them to set the collection, datetime range, bounding box, and asset.

If no asset is specfied the description is a complex source composed of multiple subdatasets ('GDAL' terminology).
In a 'STAC' context "asset" and "subdataset" are synonomous for 'GDAL'.
}
\examples{
mpc()

## we can be more general than GDAL, and say do a query to get Parquet for MS Buildings

mpc(collection = "ms-buildings", stacit = FALSE, bbox = "140,-45,145,-30",
                                 datetime = as.Date("2000-01-01"))
## read that url with jsonlite and investigate $features$assets$data$href
## but it's all abfs:// azure special links
}
