% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{read_banner_course}
\alias{read_banner_course}
\title{Combine class BannerWeb class lists into one}
\usage{
read_banner_course(path)
}
\arguments{
\item{path}{character vector of paths to XLS files exported from BannerWeb}
}
\description{
Combine class BannerWeb class lists into one
}
\examples{
\dontrun{
dir <- "~/Dropbox/SDS/students/enrollments/bannerweb_exports"
docs <- list.files(dir, pattern = "291", full.names = TRUE)
my_list <- read_banner_course(docs)
dim(my_list)
write.csv(my_list, file = "my_class.csv", row.names = FALSE)
}
}
