% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{read_major_report}
\alias{read_major_report}
\title{Read Major/Minor report}
\usage{
read_major_report(path)
}
\arguments{
\item{path}{path to Major/Minor report}
}
\value{
a tibble
}
\description{
Read Major/Minor report
}
\examples{
\dontrun{
sds <- read_major_report("/tmp/SDS_Student_MajorsMinor_Rpt_05 14 2019.pdf")
nrow(sds)
}
}
