% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{rema}
\alias{rema}
\alias{rema_v2}
\title{REMA elevation}
\usage{
rema()

rema_v2()
}
\value{
data source name for REMA v2
}
\description{
Reference Elevation Model of Antarctica, the 2m mosaic as a VRT with
rendered overviews (see github.com/mdsumner/rema-ovr).
}
\examples{
rema()
}
