% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{render_rmd}
\alias{render_rmd}
\title{Render all of the markdown files in an assigment}
\usage{
render_rmd(from, ...)
}
\arguments{
\item{from}{directory of Moodle assignment}

\item{...}{currently ignored}
}
\description{
Render all of the markdown files in an assigment
}
\examples{
\dontrun{
my_dir <- "~/Dropbox/git/sds192/student_info/test"
if (require(dplyr)) {
  my_dir \%>\%
    render_rmd()
}
}

}
