% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_smith_c}
\alias{scale_color_smith_c}
\alias{scale_color_smith_cd}
\alias{scale_color_smith_d}
\title{Smith College color palettes}
\usage{
scale_color_smith_c(reverse = FALSE, ...)

scale_color_smith_cd(reverse = FALSE, ...)

scale_color_smith_d(reverse = FALSE, ...)
}
\arguments{
\item{reverse}{Reverse the order of colors?}

\item{...}{arguments passed to methods}
}
\description{
Smith College color palettes
}
\examples{

if (require(ggplot2)) {
  ggplot(mtcars, aes(x = disp, y = mpg)) +
    geom_point(aes(color = mpg)) +
    scale_color_smith_c(reverse = TRUE)

  ggplot(mtcars, aes(x = disp, y = mpg)) +
    geom_point(aes(color = mpg)) +
    scale_color_smith_cd(midpoint = 20)

  ggplot(mtcars, aes(x = disp, y = mpg)) +
    geom_point(aes(color = factor(am))) +
    scale_color_smith_d()
}
}
\references{
\url{https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}
}
