% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_c.R
\name{Score_C_Function}
\alias{Score_C_Function}
\title{Score c function}
\usage{
Score_C_Function(attrib, clin)
}
\arguments{
\item{attrib}{Matrix of atrribute values with attribute names as row names and sample names as column names.}

\item{clin}{data.frame of clinical data with dummy variables for categorical variables and with sample names as row names and variable names as column names.}
}
\value{
Vector of scores with attribute names as names. Higher is better.
}
\description{
This function calculates the clinical independence score based on attribute values and clinical data.
}
\examples{
Score2Function()
}
\keyword{feature}
\keyword{selection}
