% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_s.R
\name{Score_S_Function}
\alias{Score_S_Function}
\title{Score s function}
\usage{
Score_S_Function(attrib, surv)
}
\arguments{
\item{attrib}{Matrix of atrribute values with attribute names as row names and sample names as column names.}

\item{surv}{data.frame of survival invormation with sample names as row names and column names 'overall.survival.indicator' and 'OVERALL.SURVIVAL'.}
}
\value{
Vector of scores with attribute names as names. Higher is better.
}
\description{
This function calculates the survival distance score based on attribute values and survival data.
}
\examples{
Score1Function()
}
\keyword{feature}
\keyword{selection}
