% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_m.R
\name{scoreCombination}
\alias{scoreCombination}
\title{Score combination}
\usage{
scoreCombination(score1, score2, weight)
}
\arguments{
\item{score1}{Vector of survival distance scores.}

\item{score2}{Vector of clinical independence scores.}

\item{weight}{Weight to apply to score 2.}
}
\value{
Vector of scores with attribute names as names. Higher is better.
}
\description{
This function combines the survival distance score (score1) and the clinical independence score (score2) with weighted standardization.
}
\examples{
scoreCombination()
}
\keyword{feature}
\keyword{selection}
