% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaice_cdr.R
\name{seaice_cdr}
\alias{seaice_cdr}
\title{Sea ice CDR (climate data record) url}
\usage{
seaice_cdr(date, hemisphere = c("south", "north"), vsi = TRUE)
}
\arguments{
\item{date}{date or or string YYYY-mm-dd}

\item{hemisphere}{north or south}

\item{vsi}{in GDAL url form}
}
\value{
string with path to CDR sea ice netcdf file
}
\description{
The first subdataset 'cdr_seaice_conc' is the one you want, but also included
are "cdr_seaice_conc_interp_spatial_flag", "cdr_seaice_conc_interp_temporal_flag",
"cdr_seaice_conc_qa_flag", "cdr_seaice_conc_stdev", "raw_bt_seaice_conc", "raw_nt_seaice_conc" and
"surface_type_mask"
}
\details{
I've had mixed success setting subdataset and having these return with the correct orientation.
}
\examples{
seaice_cdr()
}
