% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.R
\name{seir}
\alias{seir}
\title{Susceptible-(Exposed)-Infectious-Recovered}
\usage{
seir(data, model = c("SIR", "SEIR", "custom"))
}
\description{
Fit a S(E)IR model or give a ODE model.
}
\examples{
data(indiaAprilJune)
fit = seir(data = indiaAprilJune, model = "SIR")
}
