% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel_wms.R
\name{sentinel2_wms}
\alias{sentinel2_wms}
\title{Sentinel 2 WMS}
\usage{
sentinel2_wms(
  layer = c("TRUE-COLOR-S2L2A", "NDVI", "FALSE-COLOR", "FALSE-COLOR-URBAN",
    "AGRICULTURE", "BATHYMETRIC", "GEOLOGY", "MOISTURE-INDEX", "SWIR", "NATURAL-COLOR"),
  datetime = NA
)
}
\arguments{
\item{layer}{see layer options in the argument, default is "TRUE-COLOR-S2L2A"}

\item{datetime}{a valid datetime or NA for "latest"}
}
\value{
a string to a WMS
}
\description{
To use this you must have your own "INSTANCE_ID", set this in env var
"SENTINELHUB_INSTANCE_ID".
}
\examples{
sentinel2_wms()
}
