\name{sds-package}
\alias{sds-package}
\alias{sds}
\docType{package}
\title{
Functions and Exampels for ``Statistics for Data Sciences using R''
}
\description{
Functions, examples and other software related to the book ``Statistics for Data Sciences using R.''
}
\details{
\tabular{ll}{
Package: \tab sds\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-03-31\cr
License: \tab What license is it under?\cr
}
\author{
Who wrote it
Maintainer: <jharner@stat.wvu.edu>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
