% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_labs}
\alias{slack_labs}
\title{Unzips slack message exports.}
\usage{
slack_labs(dir, ...)
}
\arguments{
\item{dir}{directory}

\item{...}{currently ignored}
}
\description{
Removes previously extracted `data` directory if exists
and extracts into `data` directory.
}
\examples{
\dontrun{
slack_labs("~/Dropbox/git/sds192/student_info/SDS 192-f17 Slack export Sep 24 2017/")
}
}
