% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{tas_dem}
\alias{tas_dem}
\title{Tasmania DEM (2m)}
\usage{
tas_dem(vsicurl = TRUE)
}
\arguments{
\item{vsicurl}{prefix with vsicurl or not}
}
\value{
data source name for Tasmania 2m DEM
}
\description{
by MRT 2021
}
\examples{
tas_dem()
}
