% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_list.R
\name{term_to_banner_term}
\alias{term_to_banner_term}
\alias{banner_term_to_term}
\alias{crn_from_file}
\alias{file_refresh}
\alias{class_list_url}
\title{Utilities for dealing with BannerWeb}
\usage{
term_to_banner_term(x)

banner_term_to_term(x)

crn_from_file(dir)

file_refresh(path, domain, ...)

class_list_url(path, domain, ...)
}
\arguments{
\item{x}{Description of a term}

\item{dir}{Path to directory where BannerWeb export files are located}

\item{path}{Vector of paths to BannerWeb export files}

\item{domain}{Domain name for BannerWeb server}

\item{...}{currently ignored}
}
\description{
Utilities for dealing with BannerWeb
}
\examples{
term_to_banner_term(c("f18", "s18", "f17"))
banner_term_to_term(c("201901", "201803", "201801"))
\dontrun{
bannerweb_db <- "~/Dropbox/SDS/students/enrollments/bannerweb_exports"
crn_from_file(bannerweb_db)
}
\dontrun{
  paths <- head(fs::dir_ls(bannerweb_db))
  file_refresh(paths, "banner.myschool.edu")
}
\dontrun{
bannerweb_db <- "~/Dropbox/SDS/students/enrollments/bannerweb_exports"
paths <- head(fs::dir_ls(bannerweb_db))
class_list_url(paths, domain = "myschool.edu")
}
}
