% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{usgs_seamless}
\alias{usgs_seamless}
\title{USGS seamless DEM}
\usage{
usgs_seamless(vsicurl = TRUE)
}
\arguments{
\item{vsicurl}{if TRUE prefix /vsicurl}
}
\value{
character string, a GDAL data source name for the USGS seamless DEM
}
\description{
USGS seamless DEM
}
\examples{

usgs_seamless()

}
