% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_sirmod.R
\name{vis_seir}
\alias{vis_seir}
\title{Visualize S(E)IR Models for Disease Dynamics}
\usage{
vis_seir(model = "SIR", beta, gamma, I0, R0, times, N)
}
\description{
Visualize S(E)IR Models for Disease Dynamics
}
\examples{
vis_seir(model = "SIR", beta = 2, gamma = 1, I0 = 5, R0 = 2, time = 1:70, N = 50000)

}
