% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{xaringan}
\alias{xaringan}
\alias{html_dependency_smith}
\title{Smith xaringan slides (HTML)}
\source{
\url{https://www.smith.edu/about-smith/college-relations/visual-identity-program}
}
\usage{
xaringan(css = c("default", "default-fonts"), self_contained = FALSE,
  seal = TRUE, yolo = FALSE,
  chakra = "https://remarkjs.com/downloads/remark-latest.min.js",
  nature = list(), ...)

html_dependency_smith()
}
\arguments{
\item{css}{A vector of CSS file paths. Two default CSS files
(\file{default.css} and \file{default-fonts.css}) are provided in this
package, which was borrowed from \url{https://remarkjs.com}. If the
character vector \code{css} contains a value that does not end with
\code{.css}, it is supposed to be a built-in CSS file in this package,
e.g., for \code{css = c('default', 'extra.css')}), it means
\code{default.css} in this package and a user-provided \code{extra.css}. To
find out all built-in CSS files, use \code{xaringan:::list_css()}.}

\item{self_contained}{Whether to produce a self-contained HTML file.}

\item{seal}{Whether to generate a title slide automatically using the YAML
metadata of the R Markdown document (if \code{FALSE}, you should write the
title slide by yourself).}

\item{yolo}{Whether to insert the
\href{{https://kbroman.wordpress.com/2014/08/28/the-mustache-photo/}{Mustache
 Karl (TM)}} randomly in the slides. \code{TRUE} means insert his picture on
one slide, and if you want him to be on multiple slides, set \code{yolo} to
a positive integer or a percentage (e.g. 0.3 means 30\% of your slides will
be the Mustache Karl). Alternatively, \code{yolo} can also be a list of the
form \code{list(times = n, img = path)}: \code{n} is the number of times to
show an image, and \code{path} is the path to an image (by default, it is
Karl).}

\item{chakra}{A path to the remark.js library (can be either local or
remote).}

\item{nature}{(Nature transformation) A list of configurations to be passed
to \code{remark.create()}, e.g. \code{list(ratio = '16:9', navigation =
list(click = TRUE))}; see
\url{https://github.com/gnab/remark/wiki/Configuration}. Besides the
options provided by remark.js, you can also set \code{autoplay} to a number
(the number of milliseconds) so the slides will be played every
\code{autoplay} milliseconds. You can also set \code{countdown} to a number
(the number of milliseconds) to include a countdown timer on each slide. If
using \code{autoplay}, you can optionally set \code{countdown} to
\code{TRUE} to include a countdown equal to \code{autoplay}. To alter the
set of classes applied to the title slide, you can optionally set
\code{titleSlideClass} to a vector of classes; the default is
\code{c("center", "middle", "inverse")}.}

\item{...}{For \code{tsukuyomi()}, arguments passed to \code{moon_reader()};
for \code{moon_reader()}, arguments passed to
\code{rmarkdown::\link{html_document}()}.}
}
\description{
Write a presentation in R Markdown using Smith College SDS theming
}
\details{
This function is a thin wrapper around \code{\link[xaringan]{moon_reader}}
}
\examples{
\dontrun{
# simple invocation
rmarkdown::render("input.Rmd", "sds::xaringan")
}


}
\seealso{
\code{\link[xaringan]{moon_reader}}
}
