% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_list.R
\name{class_list}
\alias{class_list}
\title{Combine class BannerWeb class lists into one}
\usage{
class_list(xls)
}
\arguments{
\item{xls}{character vector of paths to XLS files exported from BannerWeb}
}
\description{
Combine class BannerWeb class lists into one
}
\examples{
\dontrun{
docs <- c("~/Dropbox/sds220-f16/student_info/ClassList_20218.xls",
          "~/Dropbox/sds220-f16/student_info/ClassList_20902.xls")
my_list <- class_list(docs)
dim(my_list)
write.csv(my_list, file = "my_class.csv", row.names = FALSE)
}
}

