% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{class_notes}
\alias{class_notes}
\title{Class notes format (PDF)}
\usage{
class_notes(fig_width = 10, fig_height = 2.5, fig_crop = TRUE,
  dev = "pdf", highlight = "default", keep_tex = FALSE,
  includes = NULL, md_extensions = NULL, pandoc_args = NULL)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default height (in inches) for figures}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures}

\item{dev}{Graphics device to use for figure output (defaults to pdf)}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link{includes}} function).}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}

\item{pandoc_args}{Additional command line options to pass to pandoc}
}
\description{
Class notes format (PDF)
}
\examples{
class_notes()

}
