% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{dea_250m_dem}
\alias{dea_250m_dem}
\title{DEA 250m dem (superseded)}
\usage{
dea_250m_dem(vsi = TRUE)
}
\arguments{
\item{vsi}{include the 'vsicurl' prefix (\code{TRUE} is default)}
}
\value{
errors; the source is superseded (see Description)
}
\description{
Australian Bathymetry and Topography 2023 250m MSL 'COG' (AusSeabed). This
source is dead: the 2023 S3 object was removed when it was superseded by
AusBathyTopo 250m 2024 (doi:10.26186/150050). Calling this function errors
with that successor information.
}
