% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{dea_250m_dem}
\alias{dea_250m_dem}
\title{DEA 250m dem}
\usage{
dea_250m_dem(vsi = TRUE)
}
\arguments{
\item{vsi}{include the 'vsicurl' prefix (\code{TRUE} is default)}
}
\value{
character string, URL to online raster
}
\description{
Australian Bathymetry and Topography 2023 250m MSL 'COG' (AusSeabed).
}
