% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{dsn-sources}
\alias{dsn-sources}
\alias{wms_arcgis_mapserver_ESRI.WorldImagery_tms}
\alias{wms_bluemarble_s3_tms}
\alias{wms_googlehybrid_tms}
\alias{wms_virtualearth}
\alias{wms_ESA_worldcover_2020_tms}
\alias{wms_mapbox_satellite}
\alias{wms_amazon_elevation}
\alias{wms_mapbox_terrain}
\alias{wms_openstreetmap_tms}
\alias{wms_googleterrainstreets_tms}
\alias{wms_virtualearth_street}
\alias{wms_arcgis_mapserver_tms}
\alias{nasadem}
\alias{cop90}
\alias{cop30}
\alias{srtm15}
\title{Imagery online sources}
\usage{
wms_arcgis_mapserver_ESRI.WorldImagery_tms()

wms_bluemarble_s3_tms()

wms_googlehybrid_tms()

wms_virtualearth()

wms_ESA_worldcover_2020_tms()

wms_mapbox_satellite()

wms_amazon_elevation()

wms_mapbox_terrain()

wms_openstreetmap_tms()

wms_googleterrainstreets_tms()

wms_virtualearth_street()

wms_arcgis_mapserver_tms()

nasadem()

cop90()

cop30()

srtm15()
}
\value{
character string, a GDAL data source name (a 'GDAL_WMS' XML
description or a '/vsicurl/' URL)
}
\description{
Raster and imagery online
}
\examples{
cop30()
wms_openstreetmap_tms()
}
