% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{gebco}
\alias{gebco}
\alias{gebco22}
\alias{gebco21}
\alias{gebco19}
\alias{gebco23_bedrock}
\alias{gebco23}
\title{GEBCO source dsn}
\usage{
gebco(vsi = TRUE)

gebco21(vsi = TRUE)

gebco23_bedrock(vsi = TRUE)

gebco23(vsi = TRUE)

gebco22(vsi = TRUE)

gebco19(vsi = TRUE)
}
\arguments{
\item{vsi}{include the 'vsicurl' prefix (\code{TRUE} is default)}
}
\value{
character string, URL to online GeoTIFF
}
\description{
A data source name to the GEBCO  elevation 'COG' GeoTIFF.
}
\details{
GEBCO 2022 is created and hosted by Philippe Massicotte.

GEBCO 2019 and 2021 created and hosted by the Australian Antarctic Division.

See note about which forms of the bedrock vs ice surface are available. Generally we use the ice surface form, because that is what encountered while navigating the surface of the Earth. But, the bedrock is of course also of interest.
}
\section{warning}{
 please note that \code{gebco21()}, \code{gebco19()} return the \emph{ice surface} form, while \code{gebco22()} returns the bedrock form. With \code{gebco23()} and \code{gebco23_bedrock()} these are now both avaiable, again thanks to Philippe Massicotte.
}

\examples{
gebco()
}
