% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{gedtm30}
\alias{gedtm30}
\alias{gedtm30_sources}
\title{GEDTDM global 1-arc second (30m) DEM}
\usage{
gedtm30()

gedtm30_sources()
}
\description{
Global Ensemble Digital Terrain Model 30m (GEDTM30)
https://zenodo.org/records/15490367
}
\examples{
gedtm30()
}
