% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{ghrsst}
\alias{ghrsst}
\title{GHRSST files, GeoTIFFs on source.coop}
\usage{
ghrsst(vsi = TRUE)
}
\value{
dataframe of source,date
}
\description{
GHRSST files, GeoTIFFs on source.coop
}
\examples{
files <- ghrsst()
tail(files$source)
}
