% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listmap.R
\name{list_address}
\alias{list_address}
\title{Title}
\usage{
list_address(address)
}
\value{
string URI for List service
}
\description{
Title
}
\examples{
x <- list_address(c(2862, "LYELL",  "HAYES"))
#vals <- jsonlite::fromJSON(readr::read_file ( x), simplifyVector = F)$features[[1]]$attributes
#plot(v <- vect(list_parcel(vals$PID)))
#Sys.setenv("GDAL_DISABLE_READDIR_ON_OPEN"="TRUE")
#dem <- rast("/vsicurl/https://s3.us-west-2.amazonaws.com/us-west-2.opendata.source.coop/alexgleith/tasmania-dem-2m/Tasmania_Statewide_2m_DEM_14-08-2021.tif")
#plotRGB(project(rast(ortho), rast(v, res = .1), by_util = TRUE))
#plot(v, add = T)
}
