% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsidc_seaice.R
\name{nsidc_seaice}
\alias{nsidc_seaice}
\title{Data sources for NOAA NSIDC GeoTIFF images for sea ice concentration and extent.}
\usage{
nsidc_seaice(
  date,
  hemisphere = c("south", "north"),
  temporal = c("daily", "monthly"),
  varname = c("concentration", "extent"),
  vsi = TRUE
)
}
\arguments{
\item{date}{date in POSIXct or Date format, see Details}

\item{hemisphere}{hemisphere of data source, north or south}

\item{temporal}{temporal resolution of data source, daily or monthly}

\item{varname}{variable name to obtain, concentration or extent (note this is a colour image)}

\item{vsi}{prepend DSN with "/vsicurl/" or not}
}
\value{
character string, GDAL readable DSN to a colour palette GeoTIFF (see args 'vsi' and others for options)
}
\description{
Data is colour-palette GeoTIFF in polar stereographic raster format.
}
\details{
The availability of the date is not checked currently, the minimum date is 1978-10-26. Please note that these
data were only every two days from date, until a later time (FIXME get these details).
}
\examples{
nsidc_seaice("2023-06-27")
#readBin(con <- url(nsidc_seaice("2023-06-27", vsi = FALSE), open = "rb"), "raw"); close(con)
}
