% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds-package.R
\docType{package}
\name{sds-package}
\alias{sds}
\alias{sds-package}
\title{sds: Spatial Data Sources}
\description{
Functions for descriptions of spatial data sources that can be read directly with or without download. Most sources are readable by 'GDAL' software. There are functions for imagery, elevation, and a variety of easily accessible data streams including image tile servers, links to online files, and geographic data services.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/sds}
  \item Report bugs at \url{https://github.com/hypertidy/sds/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
