% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds-package.R
\docType{package}
\name{sds-package}
\alias{sds}
\alias{sds-package}
\title{sds: Statistics for Data Science}
\description{
Statistics prerequisites for data science courses, written as Quarto fragments that host sites include through a git submodule.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Morrison-Lab/sds}
  \item Report bugs at \url{https://github.com/Morrison-Lab/sds/issues}
}

}
\author{
\strong{Maintainer}: Ezra Morrison \email{d.ezra.morrison@gmail.com}

Authors:
\itemize{
  \item Ezra Morrison \email{d.ezra.morrison@gmail.com}
}

}
\keyword{internal}
