% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.R
\name{seir}
\alias{seir}
\title{Susceptible-(Exposed)-Infectious-Recovered}
\usage{
seir(data, model = c("SIR", "SEIR", "custom"), N, start = NULL)
}
\description{
Fit a S(E)IR model or give a ODE model.
}
\examples{
data(indiaAprilJune)
N = 1366e6
fit = seir(data = indiaAprilJune, N = N, model = "SIR")
fit$coef
plot.seirMod(fit)
plot.seirMod(fit, CI = TRUE)

}
