% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_slack.R
\name{slack_labs}
\alias{slack_labs}
\title{Unzips slack message exports.}
\usage{
slack_labs(dir, ...)
}
\arguments{
\item{channel_name}{String with exact group name}

\item{date}{Date in the following format: Sep 2 2017}
}
\description{
Removes previously extracted `data` directory if exists
and extracts into `data` directory.
}
\examples{
slack_labs("~/Dropbox/git/sds192/student_info/SDS 192-f17 Slack export Sep 24 2017/")
}
