% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacit.R
\name{stacit}
\alias{stacit}
\title{Stacit query generator}
\usage{
stacit(
  extent,
  date = "",
  collections = "sentinel-2-c1-l2a",
  provider = c("https://earth-search.aws.element84.com/v1/search",
    "https://planetarycomputer.microsoft.com/api/stac/v1/search"),
  asset = c("visual"),
  gdal_stacit = FALSE,
  limit = 1000
)
}
\arguments{
\item{extent}{numeric longlat xmin,xmax,ymin,ymax OR character MGRS code (see details)}

\item{date}{character/Date/POSIXt length 1 or 2 (see details)}

\item{gdal_stacit}{}
}
\description{
STAC url query, as simple as extent and date(time).
}
\details{
Extent can be an 100km grid square MGRS code, or a number extent for longlat.

Date can be a single character "2025", "2025-01", "2025-01-01" or a pair of these for the
obvious range between dates or implied interval (day, month, year).
}
\examples{
stacit(c(140, 145, -43, -40))
stacit("43DFD", date = "2025-01")
}
