% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacit.R
\name{stacit}
\alias{stacit}
\title{Stacit query generator}
\usage{
stacit(
  extent,
  date = Sys.Date() - 1,
  collections = "sentinel-2-c1-l2a",
  provider = c("https://earth-search.aws.element84.com/v1/search",
    "https://planetarycomputer.microsoft.com/api/stac/v1/search"),
  asset = c("visual"),
  gdal_stacit = FALSE,
  limit = 1000
)
}
\arguments{
\item{gdal_stacit}{}
}
\description{
Stacit query generator
}
\examples{
stacit(c(140, 145, -43, -40))
}
